oauth flow starting to work. next up: save tokens to db

+++ client/build.sh
@@ -1,13 +1,25 @@
 #!/bin/bash
+THISDIR=$(cd $(dirname "$0"); pwd) #this script's directory
+
+pushd .
+
+cd $THISDIR
+
 clear;
+
 ./clean.sh
 
-cp -v ./src/*.html ./dist/
-cp -v ./src/*.css ./dist/
-cp -v ./src/*.js ./dist/
-
-echo "Running rollup..."
 rollup --config
-echo "Running rollup complete."
-rm -rfd ../server/data/public/*
-cp -v ./dist/* ../server/data/public/
+if [ $? -eq 0 ] ; then
+
+  cp -v ./src/*.html ./dist/
+  cp -v ./src/*.css ./dist/
+  cp -v ./src/*.js ./dist/
+
+  rm -rfd ../server/data/public/*
+  cp -v ./dist/* ../server/data/public/
+
+fi
+
+
+popd
